refactor: organize screens and address review comments
