+++ src/features/file-explorer/index.ts
@@ -1,4 +1,4 @@
-export { PullRequestPage } from "./routes/PullRequestPage";
+export { PullRequestScreen } from "./screens/PullRequestScreen";
 export { useChangedFiles } from "./hooks/useChangedFiles";
 export { buildTree, isMarkdownPath } from "./lib/buildTree";
 export type { TreeNode, FolderNode, FileLeaf } from "./lib/buildTree";
