+++ src/features/main/index.ts
@@ -1,4 +1,6 @@
-export { MainLayout, useRepoContext } from "./routes/MainLayout";
+export { MainLayout } from "./screens/MainLayout";
+export { useRepoContext } from "./hooks/useRepoContext";
+export type { RepoContext } from "./hooks/useRepoContext";
 export { SidebarShell } from "./components/SidebarShell";
 export { ThreadsPane } from "./components/ThreadsPane";
 export { PreviewSlot } from "./components/PreviewSlot";
