+++ src/features/main/screens/MainLayout.tsx
@@ -1,10 +1,10 @@
-import { Outlet, useLocation, useOutletContext, useParams } from "react-router-dom";
+import { Outlet, useLocation, useParams } from "react-router-dom";
 import { AppHeader } from "../components/AppHeader";
 import { Rail } from "../components/Rail";
+import type { RepoContext } from "../hooks/useRepoContext";
 
-interface RepoCtx {
-  owner: string;
-  repo: string;
+interface RepoRouteState {
+  branch?: string | null;
 }
 
 export function MainLayout() {
@@ -12,20 +12,17 @@ export function MainLayout() {
   const owner = params.owner ?? "";
   const repo = params.repo ?? "";
   const location = useLocation();
+  const routeState = location.state as RepoRouteState | null;
   const prMatch = location.pathname.match(/\/pulls\/(\d+)/);
   const prNumber = prMatch ? Number(prMatch[1]) : undefined;
 
   return (
     <div className="flex h-screen w-screen flex-col overflow-hidden bg-[hsl(var(--background))] text-[hsl(var(--foreground))]">
-      <AppHeader owner={owner} repo={repo} prNumber={prNumber} />
+      <AppHeader owner={owner} repo={repo} prNumber={prNumber} branch={routeState?.branch} />
       <div className="flex min-h-0 flex-1">
         <Rail />
-        <Outlet context={{ owner, repo } satisfies RepoCtx} />
+        <Outlet context={{ owner, repo } satisfies RepoContext} />
       </div>
     </div>
   );
-}
-
-export function useRepoContext() {
-  return useOutletContext<RepoCtx>();
 }
