+++ src/features/onboarding/index.ts
@@ -1,1 +1,1 @@
-export { SelectRepoRoute } from "./routes/SelectRepoRoute";
+export { SelectRepoScreen } from "./screens/SelectRepoScreen";
