+++ src/features/pull-requests/index.ts
@@ -1,3 +1,4 @@
-export { PullRequestListPage } from "./routes/PullRequestListPage";
+export { PullRequestListScreen } from "./screens/PullRequestListScreen";
 export { usePullRequests } from "./hooks/usePullRequests";
 export { useRepoPath } from "./hooks/useRepoPath";
+export { filterPullRequests } from "./lib/filterPullRequests";
